fix randomString not working with odd numbers
Due to the way the random characters are generated, they only generate strings of
even length (so like 2,4,6, etc.). Thus, if an odd number length is given, say 3,
then it woulld produce a string of length 4 instead. This is due to the APIs for
generating randon characters. This has been fixed by simply chopping off any
extra characters from the string.